Cleaned up shape feeders a bit. Registered many interesting more occupants with metaclass stuff. Click now snaps to nearest occupant. HookManager now properly recurses to base classes.

+++ src/occupant.hpp
@@ -16,8 +16,8 @@ class Occupant : public Object
     Occupant* next; // next occupant on this grid position
     Grid* grid;
     Pos pos;
-    short signature;
-    short id;
+    float signature;
+    int id;
 
     virtual void Interact(Creat& c);
     virtual void Update();
